+++ src/main/java/com/en_circle/slt/plugin/swank/SwankPacket.java
@@ -388,6 +388,16 @@ public class SwankPacket {
         return new SwankPacket(formatted);
     }
 
+    public static SwankPacket disassemble(String symbol, String packageName, BigInteger continuation) {
+        symbol = StringUtils.replace(symbol, "\\", "\\\\");
+        symbol = StringUtils.replace(symbol, "\"", "\\\"");
+//        packageName = StringUtils.replace(packageName, "\\", "\\\\");
+//        packageName = StringUtils.replace(packageName, "\"", "\\\"");
+
+        String formatted = String.format("(:emacs-rex (swank:disassemble-form \"'%s\") \":CL-USER\" T %s)", symbol, continuation);
+        return new SwankPacket(formatted);
+    }
+
     private int length;
     private String expressionSource;